dev-requirements.txt
# requirements
-r requirements.txt
# linting, typing and formatting
ruff
mypy
pytest
boto3-stubs[essential]